requirements.txt
pandas==2.2.3
numpy==1.26.4
loguru==0.7.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.27.1
scikit-learn==1.6.1
xgboost-cpu==3.0.2
optuna==3.6.1
mlflow==2.16.2